+++ src/main/java/com/pasterdream/pasterdreammod/world/effect/DyedreamPerfumeBuffEffect.java
@@ -1,0 +1,11 @@
+package com.pasterdream.pasterdreammod.world.effect;
+
+import net.minecraft.world.effect.MobEffect;
+import net.minecraft.world.effect.MobEffectCategory;
+
+public class DyedreamPerfumeBuffEffect extends MobEffect {
+
+    public DyedreamPerfumeBuffEffect() {
+        super(MobEffectCategory.BENEFICIAL, -1407043);
+    }
+}